add ToolCallRow with styling
